Implement LocaleController, enable temp database, move PetConnectApplication, and update dependencies"
