+++ src/main/java/se/storkforge/petconnect/controller/HomeController.java
@@ -1,5 +1,15 @@
 package se.storkforge.petconnect.controller;
 
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
 public class HomeController {
 
+    @GetMapping("/home")
+    public String home(Model model) {
+        model.addAttribute("message", "Welcome to PetConnect!");
+        return "index"; // Looks for index.html inside resources/templates/
+    }
 }
