+++ src/main/java/se/storkforge/petconnect/controller/LocaleController.java
@@ -1,0 +1,23 @@
+package se.storkforge.petconnect.controller;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.servlet.LocaleResolver;
+import org.springframework.web.servlet.support.RequestContextUtils;
+
+import java.util.Locale;
+
+@Controller
+public class LocaleController {
+
+    @GetMapping("/change-language")
+    public String changeLanguage(HttpServletRequest request, HttpServletResponse response, String lang) {
+        LocaleResolver localeResolver = RequestContextUtils.getLocaleResolver(request);
+        if (localeResolver != null) {
+            localeResolver.setLocale(request, response, new Locale(lang));
+        }
+        return "redirect:/home"; // Redirect back to home page after language change
+    }
+}
