+++ src/main/resources/application.properties
@@ -1,1 +1,8 @@
 spring.application.name=PetConnect
+# Use H2 in-memory database for testing
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.driver-class-name=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.hibernate.ddl-auto=create-drop
